Added painting with left and middle click.

+++ src/gfx/Rectangle.cpp
@@ -79,8 +79,8 @@ void Rectangle::onUpdate(glm::vec2 location, int button, int action) {
 
 }
 
-void Rectangle::setShaderDimensions(int winWidth, int winHeight) {
-    Rectangle::sp = std::make_shared<ShaderProgram>("Rectangle.vert", "Rectangle.frag", (float)winWidth, (float)winHeight);
+void Rectangle::setShaderDimensions(float winWidth, float winHeight) {
+    Rectangle::sp = std::make_shared<ShaderProgram>("Rectangle.vert", "Rectangle.frag", winWidth, winHeight);
 }
 
 RectangleBounds::RectangleBounds(float width, float height) {
